% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{full_width_chart_slide}
\alias{full_width_chart_slide}
\title{Add a full-width chart slide}
\usage{
full_width_chart_slide(
  slides,
  title,
  text = NULL,
  chart,
  source = NULL,
  footer = NULL
)
}
\arguments{
\item{slides}{An \code{rpptx} object (\code{\link[officer:read_pptx]{officer::read_pptx()}}).}

\item{title}{Character vector for the title.}

\item{text}{Optional character vector for the description above the chart.}

\item{chart}{A chart object, e.g. from \code{\link[mschart:ms_barchart]{mschart::ms_barchart()}}.}

\item{source}{Optional character vector for the source line.}

\item{footer}{Optional character vector.}
}
\value{
The modified \code{slides} object, invisibly.
}
\description{
Add a full-width chart slide
}
