% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck-package.R
\docType{package}
\name{deck-package}
\alias{deck}
\alias{deck-package}
\title{deck: Slide Builders for the Sage Mint PowerPoint Template}
\description{
A small set of helpers that wrap 'officer' to place titles, text, tables and charts onto the layouts of the bundled "Sage Mint Theme" PowerPoint template. Also provides 'flextable' helpers for preparing tables sized to the template's table area.
}
\author{
\strong{Maintainer}: Joe Crowley \email{joeac121@gmail.com}

}
\keyword{internal}
